Move string-equal supernodes to compare package
